Add script to create Reportico records in coreBOS
After to speak with MSLOKHAT here, in coreBOS forum: http://discussions.corebos.org/viewtopic.php?f=7&t=765&sid=d12fe114ba34ba9b418e4006427b8e6f
He sent me a script to read the reportico/projects/ directory to create records in coreBOS to be possible to execute the reports without create these manually.
We just add a ListView action to load this script and adapt the script for no read admin, tutorials, northwind directories and other that start with a dot, like .svn , .git ,.....

Also we add a check to verify that you are an admin user when click in this new button and read the reportname.xml to get the report title and report description to save in the corBOS records.

Thanks for your help and suggestions MSLOKHAT